refactor: Change type to double

+++ src/main/kotlin/blackjack/model/PlayerBet.kt
@@ -1,3 +1,3 @@
 package blackjack.model
 
-data class PlayerBet(var betAmount: Int = 0, var winnings: Int = 0)
+data class PlayerBet(var betAmount: Double = 0.0, var winnings: Double = 0.0)
